Adjusted provider agent to the updated Payment API
Since 'Settled' events are emitted it is no longer necessary to listen
for payments. Instead provider agent can listen only for invoice events
which inform both about acceptance and payment.

Invoice issuing was refactored to have a separate retry loop for
issuing itself and sending invoice to the requestor. This way no
'orphaned' invoices are created.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>